Add Task Refactor
